+++ packages/spacecat-shared-utils/test/index.test.js
@@ -24,10 +24,10 @@ describe('Index Exports', () => {
     'deepEqual',
     'fetch',
     'generateCSVFile',
-    'getRUMDomainKey',
     'getStoredMetrics',
     'replacePlaceholders',
     'getPrompt',
+    'getQuery',
     'hasText',
     'isArray',
     'isBoolean',
@@ -56,6 +56,9 @@ describe('Index Exports', () => {
     'stripWWW',
     'toBoolean',
     'tracingFetch',
+    'getHighFormViewsLowConversionMetrics',
+    'getHighPageViewsLowFormViewsMetrics',
+    'getHighPageViewsLowFormCtrMetrics',
   ];
 
   it('exports all expected functions', () => {